pyproject.toml
[project]
name = "kompassi"
version = "2.0.0"
description = "Event management system for volunteer-run conventions"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lippukala @ git+https://github.com/kcsry/lippukala@v2.0.0",
    "paikkala @ git+https://github.com/kcsry/paikkala@v0.2.0",
    "aiohttp[speedups]",
    "Babel",
    "bleach",
    "boto3",
    "celery",
    "crispy-bootstrap3",
    "django_csp",
    "django-bootstrap3",
    "django-cors-headers",
    "django-crispy-forms",
    "django-environ",
    "django-ipware",
    "django-localflavor",
    "django-oauth-toolkit",
    "django-redis",
    "django-storages",
    "django-stubs",
    "Django",
    "fastapi[standard-no-fastapi-cloud-cli]",
    "flake8",
    "graphene-django",
    "graphene-pydantic",
    "gunicorn",
    "ics",
    "ipdb",
    "ipython",
    "isort",
    "jinja2",
    "jsonschema",
    "jwskate",
    "markdown",
    "paramiko",
    "passlib",
    "phonenumberslite",
    "Pillow",
    "pre-commit",
    "psycopg[c,pool]",
    "pydantic[email]",
    "pypugjs",
    "pyright",
    "pytest-django",
    "python-dateutil",
    "PyYAML",
    "regex",
    "reportlab",
    "requests",
    "requests-oauthlib",
    "ruff",
    "tabulate",
    "unicodecsv",
    "uvicorn[standard]",
    "vobject",
    "weasyprint",
    "XlsxWriter",
    "zxcvbn",
]

[tool.black]
line-length = 120

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
ignore-init-module-imports = true
extend-select = [
    "B",
    "C",
    "I", # import sorting
    "PERF",
    "G010",  # .warn is deprecated
    "PGH004",  # specific noqa
    "PL",  # pylint
    "RET",
    "RUF005",  # avoid iterable concatenation
    "RUF013",  # implicit `optional` (contra type annotations)
    "RUF100",  # unused noqa
    "SIM",
    "UP",
    "S101",  # asserts
]
extend-ignore = [
    "C408",  # unnecessary dict call (these are likely stylistic)
    "C901",  # function is too complex
    "E401",  # multiple imports on one line (handled by I)
    "E731",  # do not assign a lambda expression, use a def
    "PLC0415",  # import must be at the top of the file (but we have circulars)
    "PLR0911",  # too many returns
    "PLR0912",  # too many branches
    "PLR0913",  # too many args
    "PLR0915",  # too many statements
    "PLW2901",  # for loop variable overwritten (too many false positives)
    "RET505",  # Unnecessary `else` after `return` statement
    "RET507",  # Unnecessary `elif` after `continue` statement
    "SIM105",  # contextlib.suppress has a performance cost
    "SIM108",  # ternary operators can be less readable
    "PERF401", # manual-list-comprehension
    "PLR2004",  # I will not, for the love of God, write AMOUNT_OF_MONTHS_IN_YEAR = 12
    "UP038", # isinstance(foo, X | Y) rule is deprecated
]

[tool.ruff.lint.extend-per-file-ignores]
"**/__init__.py" = [
    "F401",  # imported but unused
]
"events/**/setup_*.py" = [
    "B007",  # loop control variable not used
    "F841",  # local variable is assigned to but never used
    "PERF",  # setups are one-off, PERF is not relevant
]
"**/test*.py" = [
    "S101",  # asserts are ok in tests
]
"events/*201*/*.py" = [
    "S101",  # asserts are ok in legacy events
]
"events/*202[0123]/*.py" = [
    "S101",  # asserts are ok in legacy events
]
"**/management/**.py" = [
    "S101",  # asserts are ok in one-off management commands
]
"kompassi/wsgi.py" = [
    "E402",  # vscode for some reason strips this off as a noqa
]

[tool.pytest.ini_options]
addopts = "--verbose --reuse-db -m 'not integration_test'"
DJANGO_SETTINGS_MODULE = "kompassi.settings"
python_files = "tests.py test_*.py *_tests.py"
markers = [
    "integration_test",
]
